Add RPC::MessageSocket class
From the class docs:
A MessageSocket is a message-oriented layer on top of a TCP connection.
It sends and receives discrete chunks of data identified by opaque IDs.
Higher layers can use this to build an RPC framework, both on the client
side and on the server side.

Some of the code is based loosely on RAMCloud's TCPTransport.