Add a real RPC layer on top of opaque RPCs
This will make it possible to support multiple services on a single TCP
port.

Unfortunately, this commit grew out of control in size. It introduced a
new layer to the RPC system and ported existing code over to it. In
addition to changing a bunch of actual code, this especially caused a
lot of churn in tests.