Remove RaftLog::getTerm(), fix unsafe getEntry calls
This was making me a bit nervous, since it just returned 0 when the
entry you asked for wasn't part of your log. With snapshots, that gets a
little scary. The alternative now is to call getEntry(index).term(),
which will throw an exception if the entry isn't found.

All RaftLog::getEntry() calls should now be safe even if a prefix is
truncated after a snapshot.