Add last included term to snapshots
Before the code assumed the log always overlapped with the snapshot
in the last entry, so all terms could come from the log. Unfortunately
that doesn't really make sense, since for example with a new follower,
you have to send a snapshot first, then send log entries (to satisfy the
consistency check).

So instead, snapshot headers now include the term of the last entry they
cover, and this term is used throughout the code.