Add ConfigurationManager class
This class basically shadows the configurations found in the log and
latest snapshot, and sets the current configuration accordingly. Making
this a separate class cleans up this logic that was previously scattered
throughout the consensus code, makes it testable and more efficient, and
should make it easier to truncate prefixes from the log.