Fix wedge in upToDateLeader() when log prefixes are discarded
The problem occurred before if a snapshot was taken covering the entire
log, then the entire log was discarded. In that case, because the method
didn't consult lastSnapshotTerm, it couldn't tell that the term of the
commitIndex entry was its current term, so upToDateLeader was stuck in
an infinite sleep.