Fix invalid memory access after socket disconnect
Fix #1: valgrind error in RPCClientServerTest.timeout test

The MessageSocket code attempted to read, then write the socket. If it
was disconnected, it was unsafe to do both, since the socket object
could get destroyed in between. Now it just handles one event and
returns to the event loop.

Thank you valgrind :)