Fix disruptive servers after downsizing a cluster
Now RequestVotes are rejected for a base election timeout after hearing
a heartbeat, as described in the Raft paper. Also, servers now start
elections even when they're not part of their current configuration; the
code before had an explicit check to prevent this, but it's needed (see
thesis).

Fix #56: "fix disruptive servers after downsizing a cluster"