WIP - Ugly mess of stuff used for Diego's dissertation
This is a big rats' nest of changes, none of which are complete or
tested enough to go into master quite yet. I used something close to
this to produce the results in my dissertation, modulo a few flags and
options.

I don't expect anyone to be able to run this code, nor will I support
it. I'm uploading now so that I can later take the good bits and pieces
from here, clean them up, and commit to master in a more sane form. It
was never meant to be public, but it doesn't seem particularly
incriminating.

These changes are (c) 2014 Stanford University. Since I've graduated,
this will likely be my last commit on their behalf.