Fix constructing ClientSession with invalid address
On some hosts (such as RHEL6), connect() returns OK for a 0-length
sockaddr, which would cause issues later on when the socket was, in
fact, not connected. This adds a check to make sure it doesn't call
connect() that way.

Fix #66: crash at LogCabin::RPC::MessageSocket::read
(RPC/MessageSocket.cc:292)