Add timeouts to waiting for RPC::ClientRPCs
This exposes a timeout all the way up from the ClientSession condition
variable up to ClientRPC, so that one day clients will be able to time
out their requests. Note that the client library does not yet make use
of these timeouts; that will be in a subsequent commit.

Towards #69: add timeouts to client calls