Add ServerStats structure on servers
This holds state, metrics, etc about a server that may be useful for
diagnostics. It's currently only available on each server, and printed
out as a NOTICE to the debug log on SIGUSR1 and periodically (every
minute by default). The plan is for this same structure to be available
over to (special?) clients over RPC in the future.

Currently, the stats include most of the easily available Raft state.
This is probably pretty valuable information already, but I expect it to
evolve over time to contain more useful information.

Towards #77: tool to extract stats from servers