Fix read-only query latency for large heartbeat intervals
Read-only queries wait for a round of heartbeats to come back from a
majority of the cluster. They were meant to schedule a round of
heartbeats right away, but the code was missing a broadcast to a
condition variable, causing the read-only queries to wait for the
regularly scheduled heartbeats. This isn't so terrible with the default
heartbeat interval of 75ms, but it's not ok for higher intervals.