Add Storage/Tool executable to dump out contents of log and snapshot
This gets installed as logcabin-storage.

This introduces file locking in the LogCabin daemon for the storage
directory, so that the daemon and this tool cannot run at the same time
(this command will open the log in the normal way, which modifies the
log's storage).

This also adds various utilities:

- Core::StringUtil::join() to create strings like '1,2,3' from the
  vector {"1", "2", "3"}.
- Core::StringUtil::flags() to print things like LOCK_EX|LOCK_NB in a
  user-friendly way.
- Core::Util::Finally to call the given std::function when an object
  goes out of scope.