Add new Storage::Layout class
Storage::Layout encapsulates how LogCabin lays out the filesystem, and
it handles locking of the storage directory.

This changes the file locking from taking an flock() on the server's
directory to taking an flock() on a file called "lock" in that
directory. Locking the directory itself doesn't seem to work across NFS
clients (at least on RHEL6), whereas locking a file does.