Update README and change config file to take listen addreses
With this change, you no longer specify the list of servers in the
config file. You now specify just the address(es) you want the server to
listen on. That means that each server must have its own config file
(which I think is clearer, anwyay).

Updates README with this change and gets rid of its use of DNS.

Updates scripts/smoketest.py to stop using DNS.

Towards #63: update README and scripts to use lists of addresses instead
of DNS. (The Reconfigure binary should still be reworked.)

Close #85: After executing, "Reconfigure", leader still not syncing up
with the followers. This was blocked on the README update.