Add a couple of fields to Raft.proto that might be useful later
RequestVote now returns whether the caller's log is up-to-date, and
AppendEntries now returns the recipient's last log index. These aren't
currently used, but they're needed for some extensions to Raft that we
may want to implement later.

Also renamed the opcode for INSTALL_SNAPSHOT, which was still set to the
old name of APPEND_SNAPSHOT_CHUNK.

Towards #99: versioning