Add low-level version message type
This can be used if we ever need to change the low-level framing
protocol used by MessageSocket. The client would send a version message
request using version 1 of the protocol, and the server would reply
(using version 1) with the maximum version it supports. This could be
done in a blocking manner as the first message on a connection,
or it could be done in parallel (depending on whether the upgrade
is required or optional).

Since there's only version 1 for now, clients do not send requests of
this type for now.