Close #67: add cluster-wide UUID
From sample.conf docs for clusterUUID:

An opaque string used to prevent accidental communication across LogCabin
clusters. If set, this string will be checked when creating each
client-to-server and server-to-server session. If the recipient's has a cluster
UUID that does not match the UUID in the request, the connection will not be
closed with an error.

The default (empty string) allows communication with any client and server of
any cluster initially. Once this process discovers a cluster UUID, it will
insist on using it when creating future sessions.

Set this to a unique string for each cluster to avoid accidental
communication across LogCabin clusters (which could have disastrous effects
on your data).