Add basic functionality to update state machine
Now servers report back which state machine versions they support
(piggy-backed on AppendEntries responses). Leaders will append
AdvanceStateMachineVersion commands to the log based on this
information, so that state machines will all update their behavior at
the same log position in the future.

This is most of #125, but state machines currently panic if they receive
an unsupported command, which is probably not what we want to happen.

Towards #125: need way to transition state machine code