Keep history of state machine version updates
LogCabin's state machine changes its state as one step, and later
a ClientService worker thread may call back into the state machine to
retrieve a client response. When retrieving the result, the state
machine needs to know what version it was running at the time it
processed the command. For example, maybe the version number was lower
then, and it ignored the command entirely then. Thus, it's not
sufficient for the state machine to track the current version it's
running.

This commit changes the state machine to maintain the history of all
version upgrades (log index and version number pairs). This is
serialized into and loaded from snapshots. This history is kept for all
time; its space usage is small as it is bounded by the number of state
machine version changes in the code.

Since a 2-byte integer is no longer sufficient to serialize the current
running version into a snapshot, this changes the state machine snapshot
format to use a protobuf for the version update history. This is merged
with the existing sessions protobuf as what's now called the snapshot's
state machine header, and that explains much of the code churn in this
commit.

Towards #125: need way to transition state machine code