Leave signals blocked on server shutdown
This works around a specific problem: when running the servers under
valgrind through cluster.py, the servers would receive SIGTERM, start to
shut down, then the instant the SIGTERM signal was unmasked, the server
would appear to exit with a 0 status, yet it wouldn't finish the
shutdown process. I couldn't reproduce this outside of cluster.py. As
there's no reason to unblock the signals before exiting the daemon, this
seems like the safer bet for now.