Turn on -fno-strict-overflow
This makes signed integer overflow be well-defined, so that optimizing
compilers don't take it out. Affects at least the code commented in
Core/Time.cc and any number of other unknown places.

Using the flag under both Clang and GCC. It was Clang with -O2 that
optimized away the check in Core/Time.cc.

Towards #9: port to clang