critical hotfix for #160: Packing entries into AppendEntries requests is broken
This fixes a safety issue that could lead to data corruption. When
adding entries to AppendEntries requests that are approaching 1MB in
cumulative size, the leader could mistakenly add a noncontiguous entry
from later in the log. See the GitHub issue for details:
https://github.com/logcabin/logcabin/issues/160

More safeguards and testing will follow, but this is an obvious and
important fix that is important to get out there. This patch should be
trivial to backport to any prior commit of LogCabin.