Augment failovertest.py to start multiple clients
This allows you to do things like running the FailOverTest client
alongside the Benchmark client in a convenient way.

I was using it like this (with a large snapshotMinLogSize and maxThreads
defined in smoketest.conf, YMMV):

   ./scripts/failovertest.py \
      --killinterval=120 --launchdelay=60 --timeout=6000 \
      --client='build/Examples/FailoverTest --timeout=30s' \
      --client='build/Examples/Benchmark --threads=100 --size=100 --timeout=1000000 --writes=1000000000000'

This is enough on #168 for my purposes for now.

Close #168: create new system test with server that is far behind