Improve performance of packing entries into AppendEntries requests
Before, it possible for packing up the requests to take so long as to
cause availability and performance problems. The number of calls to
calculate the ProtoBuf's size have been greatly reduced now, and the
number of entries in any given request is also limited to 5000.

Close #161: Packing entries into AppendEntries requests is slow