Reduce wasted bandwidth when slow followers are down
Before, if a leader knew that a follower was far behind, it would send
large AppendEntries batches to that follower, even if RPCs to that
follower were failing. This makes it send empty AppendEntries RPCs
(heartbeats) instead, until getting an acknowledgment back.

Renames forceHeartbeat to suppressBulkData. The same concept applies to
InstallSnapshot RPCs, where we want to send 0 data bytes after an RPC
failure until we get an acknolwedgment.