Fix dangling pointer in MockClientImpl
The issue was that the destructor for ClientImpl calls
ExactlyOnceRPCHelper.exit() which calls into LeaderRPC. Only in case of
MockClientImpl, TreeLeaderRPC needed MockClientImpl to remain live, but
its destructor had already started. This removes the pointer that
TreeLeaderRPC had to MockClientImpl; it now keeps a reference to the
TestingCallbacks directly.

Valgrind run of unit tests is now clean again.

Close #175: Crash in unit tests using MockClientImpl