Add handler for SIGUSR2 to reopen debug log file
This will be used for logrotate.

This also adds a few methods to Core::Debug for setting a debug log
filename and reopening it. Since those might as well go into the public
API, semver requires a minor version bump, so the next release will
probably be known as v1.1.0.

Close #150: support for logrotate