Clean up signal usage in StateMachine
The snapshotting child process now unblocks all signals, since it
shouldn't be listening for them anyhow. This allows SIGTERM to be used
for signalling from the parent to the child.

Before SIGHUP was used to ask the snapshotting child to exit. Now
SIGTERM is used for a clean exit and SIGKILL is used for a watchdog
exit.

This required making signals unblockable and piping that through to the
state machine.