Rename Examples::Util::parseTime to Core::Time::parseDuration
The various client binaries used to call Examples::Util::parseTime() to
convert strings like 30ms into a number of nanoseconds. This is useful
inside LogCabin as well, so now that functionality has been moved
Core::Time::parseTime. So that clients can still access it, it is now
exposed in a new header file <LogCabin/Util.h>. The API is marked
unstable for now.