Reduce use of abort() for non-bugs
Allowing a program to throw an uncaught exception or calling PANIC each
call abort(), which creates a core file on some systems. That's not
desirable for runtime errors such as if the user input was invalid or
the port was in use, etc.

This commit adds catch blocks to many of the main functions for
certain exception types. It also introduces a macro EXIT which is like
PANIC but calls exit(1) instead of abort(), and switches a few uses to
that. There may be a few more that were missed, and finding and
converting those will have to be an ongoing process.

Close #166: catch all exceptions, then exit 1