issue #205: do not start election after removing self from cluster
a leader that has just left the cluster can win an
election. it will step down so this is just a temporary
situation but it can cause extended outages under
certain configuartions. with this change, a server
will detect that it has written itself out of the
cluster and will not run for election