issue #213: fix deadlock in LogCabin::Core:Debug at flockfile
in extensive testing with logrotate and verbose logging,
logcabin will occassionally hang at flockfile. the cause
is still a bit mysterious but there are multiple people
claiming there are issues with flockfile and signal
handlers. this commit pre-creates the log message so the
internal flockfile call can be used. in testing, the
problem has gone away. in all cases, this commit should
not break anything, so i would favor merging it