issue-219: redhat-init: override racy check for process running
the redhat provided status function checks that
the pid in a given pidfile exists. if a process
is killed or crashes without using the service
script, a new process could take the pid, falsely
indicating the process is running. this change
overrides the provided functions in favor of ones
that can also check that the command line in /proc
matches the program name